feat : checkBadge service
#27

+++ src/controllers/whitelist.controller.ts
@@ -1,10 +1,21 @@
-import { Body, Controller, Delete, Get, Param, Post } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Post,
+  Request,
+  UseGuards,
+} from '@nestjs/common';
 import {
   ApiCreatedResponse,
   ApiOperation,
   ApiParam,
   ApiTags,
 } from '@nestjs/swagger';
+import { LoginAuthGuard } from 'src/auth/guard/login.guard';
+import { UserService } from 'src/services/user/user.service';
 import { GetWhiteListResponseDto } from 'src/services/whitelist/dto/get-whitelist.dto';
 import {
   PostWhitelistRequestDto,
@@ -15,9 +26,13 @@ import { WhitelistService } from 'src/services/whitelist/whitelist.service';
 @Controller('whitelist')
 @ApiTags('Whitelist')
 export class WhitelistController {
-  constructor(private readonly whitelistService: WhitelistService) {}
+  constructor(
+    private readonly whitelistService: WhitelistService,
+    private readonly userService: UserService,
+  ) {}
 
   @Get('/')
+  @UseGuards(LoginAuthGuard)
   @ApiOperation({
     summary: 'Get Whitelist',
     description: 'Get Whitelist',
@@ -26,11 +41,15 @@ export class WhitelistController {
     description: 'Whitelist list',
     type: GetWhiteListResponseDto,
   })
-  async findAll(): Promise<GetWhiteListResponseDto> {
-    return await this.whitelistService.findAll();
+  async findAll(@Request() req): Promise<GetWhiteListResponseDto> {
+    const userId = req.user.userId;
+    if (await this.userService.checkBadgeForAuth(userId, 'admin')) {
+      return await this.whitelistService.findAll();
+    }
   }
 
   @Post('/')
+  @UseGuards(LoginAuthGuard)
   @ApiOperation({
     summary: 'Create Whitelist',
     description: 'Create Whitelist',
@@ -40,18 +59,26 @@ export class WhitelistController {
     type: PostWhitelistResponseDto,
   })
   async create(
-    @Body() request: PostWhitelistRequestDto,
+    @Request() req,
+    @Body() body: PostWhitelistRequestDto,
   ): Promise<PostWhitelistResponseDto> {
-    return await this.whitelistService.create(request);
+    const userId = req.user.userId;
+    if (await this.userService.checkBadgeForAuth(userId, 'admin')) {
+      return await this.whitelistService.create(body);
+    }
   }
 
   @Delete('/:id')
+  @UseGuards(LoginAuthGuard)
   @ApiParam({ name: 'id', description: 'Whitelist id' })
   @ApiOperation({
     summary: 'Delete Whitelist',
     description: 'Delete Whitelist',
   })
-  async delete(@Param('id') id: string): Promise<void> {
-    await this.whitelistService.delete(id);
+  async delete(@Request() req, @Param('id') id: string): Promise<void> {
+    const userId = req.user.userId;
+    if (await this.userService.checkBadgeForAuth(userId, 'admin')) {
+      await this.whitelistService.delete(id);
+    }
   }
 }
